// src/api/generated/toodlApi.zod.ts
export const userUpdatePasswordBodyNewPasswordMin = 8;
export const userUpdatePasswordBodyNewPasswordMax = 50;

export const userUpdatePasswordBodyConfirmPasswordMin = 8;
export const userUpdatePasswordBodyConfirmPasswordMax = 50;

export const userUpdatePasswordBodyOldPasswordMin = 8;
export const userUpdatePasswordBodyOldPasswordMax = 50;



export const UserUpdatePasswordBody = zod.object({
  "newPassword": zod.string().min(userUpdatePasswordBodyNewPasswordMin).max(userUpdatePasswordBodyNewPasswordMax),
  "confirmPassword": zod.string().min(userUpdatePasswordBodyConfirmPasswordMin).max(userUpdatePasswordBodyConfirmPasswordMax),
  "oldPassword": zod.string().min(userUpdatePasswordBodyOldPasswordMin).max(userUpdatePasswordBodyOldPasswordMax).optional()
})